converted the admin add activity into a fragment
