+++ app/src/main/AndroidManifest.xml
@@ -12,22 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Plannr"
         tools:targetApi="31">
-        <activity
-            android:name=".admin.adminAdd.AdminAddView"
-            android:exported="true">
 
-            <!-- delete the intent after -->
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-            <!-- delete the intent after -->
-
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
         <activity
             android:name=".MainActivity"
             android:exported="true"
