+++ app/src/main/res/layout/fragment_admin_add.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".admin.adminAdd.AdminAddView">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context=".admin.adminAdd.adminAddFragment">
 
     <TextView
         android:id="@+id/addCourseText"
